:sparkles: 기존 페이지 이름 변경
