+++ src/pages/index.ts
@@ -7,6 +7,6 @@ export const Login = lazy(() => import('./login'));
 export const Search = lazy(() => import('./search'));
 export const SelectInterest = lazy(() => import('./selectInterest'));
 export const Setting = lazy(() => import('./setting'));
-export const FindPassword = lazy(() => import('./findPassword'));
+export const GetTempPassword = lazy(() => import('./getTempPassword'));
 export const Register = lazy(() => import('./register'));
 export const ManageCategory = lazy(() => import('./manageCategory'));
